Make frequency field a `u64`
It's not an array index and is potentially pulled from an 8-byte number,
so this is more correct.

+++ src/node_types.rs
@@ -4,7 +4,7 @@ pub(crate) fn read_00<'buf>(
     buf: &'buf [u8],
     base: usize,
     _ind: usize,
-    freq: usize,
+    freq: u64,
 ) -> crate::Link<'buf> {
     crate::Link {
         ch: buf[base],
@@ -18,11 +18,11 @@ pub(crate) fn read_10<'buf>(
     buf: &'buf [u8],
     base: usize,
     ind: usize,
-    _freq: usize,
+    _freq: u64,
 ) -> crate::Link<'buf> {
     crate::Link {
         ch: buf[base + 2 * ind],
-        freq: buf[base + 2 * ind + 1] as usize,
+        freq: buf[base + 2 * ind + 1] as u64,
         buf,
         loc: None,
     }
@@ -32,12 +32,12 @@ pub(crate) fn read_11<'buf>(
     buf: &'buf [u8],
     base: usize,
     ind: usize,
-    _freq: usize,
+    _freq: u64,
 ) -> crate::Link<'buf> {
     let ofs = buf[base + 3 * ind + 2];
     crate::Link {
         ch: buf[base + 3 * ind],
-        freq: buf[base + 3 * ind + 1] as usize,
+        freq: buf[base + 3 * ind + 1] as u64,
         buf,
         loc: if ofs == u8::MAX {
             None
@@ -51,12 +51,12 @@ pub(crate) fn read_12<'buf>(
     buf: &'buf [u8],
     base: usize,
     ind: usize,
-    _freq: usize,
+    _freq: u64,
 ) -> crate::Link<'buf> {
     let ofs = LE::read_u16(&buf[base + 4 * ind + 2..]);
     crate::Link {
         ch: buf[base + 4 * ind],
-        freq: buf[base + 4 * ind + 1] as usize,
+        freq: buf[base + 4 * ind + 1] as u64,
         buf,
         loc: if ofs == u16::MAX {
             None
@@ -70,12 +70,12 @@ pub(crate) fn read_22<'buf>(
     buf: &'buf [u8],
     base: usize,
     ind: usize,
-    _freq: usize,
+    _freq: u64,
 ) -> crate::Link<'buf> {
     let ofs = LE::read_u16(&buf[base + 5 * ind + 3..]);
     crate::Link {
         ch: buf[base + 5 * ind],
-        freq: LE::read_u16(&buf[base + 5 * ind + 1..]) as usize,
+        freq: LE::read_u16(&buf[base + 5 * ind + 1..]) as u64,
         buf,
         loc: if ofs == u16::MAX {
             None
@@ -89,12 +89,12 @@ pub(crate) fn read_88<'buf>(
     buf: &'buf [u8],
     base: usize,
     ind: usize,
-    _freq: usize,
+    _freq: u64,
 ) -> crate::Link<'buf> {
     let ofs = LE::read_u64(&buf[base + 17 * ind + 9..]);
     crate::Link {
         ch: buf[base + 17 * ind],
-        freq: LE::read_u64(&buf[base + 17 * ind + 1..]) as usize,
+        freq: LE::read_u64(&buf[base + 17 * ind + 1..]) as u64,
         buf,
         loc: if ofs == u64::MAX {
             None
